+++ my-project1/src/pages/AdminNotifications.js
@@ -8,7 +8,6 @@ import { toast, ToastContainer } from "react-toastify";
 const AdminNotifications = () => {
   const [activeTab, setActiveTab] = useState("notifications");
   const [notifications, setNotifications] = useState([]);
-  const [formData, setFormData] = useState({ name: "", email: "", role: "user", password: "" });
   const [viewData, setViewData] = useState(false);
   const [showViewModal, setShowViewModal] = useState(false);
   const [searchQuery, setSearchQuery] = useState("");
@@ -36,16 +35,18 @@ const AdminNotifications = () => {
     }
   };
 
-  const handleDelete = async (id) => {
-    const apiUrl = "http://localhost:5000/api/notifications";
+  const handleDelete = async (id, name, email, message) => {
+    if (window.confirm(`Are you sure you want to delete ${message} for ${name} (${email})?`)) {
+      const apiUrl = "http://localhost:5000/api/notifications";
 
-    try {
-      await axios.delete(`${apiUrl}/${id}`);
-      fetchData();
-      toast.success("Announcement notification deleted successfully!");
-    } catch (error) {
-      console.error("Error deleting data:", error);
-      toast.error("Failed to delete announcement notification.");
+      try {
+        await axios.delete(`${apiUrl}/${id}`);
+        fetchData();
+        toast.success("Announcement notification deleted successfully!");
+      } catch (error) {
+        console.error("Error deleting data:", error);
+        toast.error("Failed to delete announcement notification.");
+      }
     }
   };
 
@@ -89,10 +90,10 @@ const AdminNotifications = () => {
             { id: 'roomNo', label: 'Room No', minWidth: 100 },
             { id: 'name', label: 'Name', minWidth: 150 },
             { id: 'email', label: 'Email', minWidth: 200 },
-              // { id: 'guests', label: 'Guests', minWidth: 100, align: 'center' },
-              // { id: 'checkInDate', label: 'Check-in Date', minWidth: 150, align: 'center' },
-              // { id: 'periodOfStay', label: 'Period of Stay', minWidth: 150, align: 'center' },
-              { id: 'message', label: 'Message', minWidth: 200, align: 'center' },
+            // { id: 'guests', label: 'Guests', minWidth: 100, align: 'center' },
+            // { id: 'checkInDate', label: 'Check-in Date', minWidth: 150, align: 'center' },
+            // { id: 'periodOfStay', label: 'Period of Stay', minWidth: 150, align: 'center' },
+            { id: 'message', label: 'Message', minWidth: 200, align: 'center' },
             { id: 'actions', label: 'Actions', minWidth: 120, align: 'center' }
           ]}
           rows={filteredNotifications.map(user => user.role == "user" && ({
@@ -106,7 +107,7 @@ const AdminNotifications = () => {
             actions: (
               <div style={{ display: "flex", gap: "15px", justifyContent: "center" }}>
                 <button onClick={() => handleView(user._id)} className="text-blue-600"><i className="fa fa-eye" aria-hidden="true"></i></button>
-                <button onClick={() => handleDelete(user._id)} className="text-red-600"><i className="fa fa-trash" aria-hidden="true"></i></button>
+                <button onClick={() => handleDelete(user._id, user.name, user.email, user.message)} className="text-red-600"><i className="fa fa-trash" aria-hidden="true"></i></button>
               </div>
             )
           }))}
